+++ SZRST.API/SZRST.API/Controllers/ReservationReportController.cs
@@ -95,7 +95,7 @@ namespace SZRST.Web.Controllers
         public int Year { get; set; }
         public int Month { get; set; }
         public int TotalReservations { get; set; }
-        public float Profit { get; set; }
+        public decimal Profit { get; set; }
 
         public string FacilityName { get; set; }
     }
